+++ src/main/java/services/StaffRepoServices.java
@@ -2,7 +2,8 @@ package services;
 
 import entities.Staff;
 import repository.StaffRepo;
-import utils.DuplicateUser;
+import exceptionHandler.DuplicateUser;
+import exceptionHandler.NoSuchId;
 
 import java.util.List;
 
@@ -12,28 +13,38 @@ public class StaffRepoServices implements BaseServices<Staff> {
 
     @Override
     public Staff add(Staff staff) {
-
-        if (staffRepo.showInfo(staff.getUsername(),Staff.class)!=null) {
-            return staffRepo.add(staff);
+        if (checkUsername(staff.getUsername())){
+            throw new DuplicateUser("username already exists!");
         }
-        else throw new DuplicateUser("username already exists");
+        return staffRepo.add(staff);
     }
 
     @Override
     public Staff remove(int id) {
 
-        Staff returnedStaff=staffRepo.showInfo(id,Staff.class);
+        if (!checkId(id)) {
+            throw new NoSuchId();
+        }
+        Staff returnedStaff=staffRepo.showInfo(Staff.class,id);
         return staffRepo.remove(returnedStaff);
     }
 
     @Override
     public void update(Staff staff) {
+
+        if (!checkId(staff.getId())){
+            throw new NoSuchId();
+        }
         staffRepo.update(staff);
     }
 
     @Override
     public Staff showInfo(int id) {
-        return staffRepo.showInfo(id,Staff.class);
+
+        if (!checkId(id)){
+            throw new NoSuchId();
+        }
+        return staffRepo.showInfo(Staff.class,id);
     }
 
     @Override
@@ -42,7 +53,43 @@ public class StaffRepoServices implements BaseServices<Staff> {
     }
 
 
+
     public Staff showInfo(String username) {
-        return null;
+
+        if (!checkUsername(username)){
+            throw new NoSuchId();
+        }
+        return staffRepo.showInfo(Staff.class,username);
+    }
+
+
+    public boolean checkUsername(String username){
+
+        boolean flag=false;
+        List<Staff> staffList=showAll();
+
+        for (Staff a:staffList){
+            if (username.equals(a.getUsername())){
+                flag=true;
+                break;
+            }
+        }
+        return flag;
+    }
+
+//checking if the id exists in order to handle null pointer exception
+//of nat having an entity with such id
+    public boolean checkId(int Id){
+
+        boolean flag=false;
+        List<Staff> staffList=showAll();
+
+        for (Staff s:staffList){
+            if (s.getId()==Id){
+                flag=true;
+                break;
+            }
+        }
+        return flag;
     }
 }
